saga and navigation structure
